requirements.txt
clint
pyaml
pyfiglet
